Added swagger doc for places API
