+++ api/src/controllers/index.ts
@@ -1,0 +1,5 @@
+import placesController from './places';
+
+export {
+    placesController,
+};
