+++ api/src/staticModels/PlaceConfirmedStatuses.ts
@@ -1,7 +1,7 @@
 class PlaceConfirmedStatuses {
     static Confirmed = 1;
 
-    static New = 0;
+    static New = 2;
 
     public static isValid(status: number): boolean {
         return status === PlaceConfirmedStatuses.Confirmed || status === PlaceConfirmedStatuses.New;
